+++ src/components/select/multi-select/multi-select.style.ts
@@ -3,9 +3,6 @@ import StyledPill from "../../pill/pill.style";
 import applyBaseTheme from "../../../style/themes/apply-base-theme";
 import InputIconToggleStyle from "../../../__internal__/input-icon-toggle/input-icon-toggle.style";
 import StyledSelect, { type StyledSelectProps } from "../select.style";
-import InputPresentationStyle from "../../../__internal__/input/input-presentation.style";
-import StyledInput from "../../../__internal__/input/input.style";
-import sizes from "../../../__internal__/input/input-sizes.style";
 
 const StyledSelectPillContainer = styled.div.attrs(applyBaseTheme)`
   display: flex;
@@ -30,16 +27,6 @@ const StyledSelectMultiSelect = styled(
     position: absolute;
     right: 0;
     height: 100%;
-  }
-
-  ${InputPresentationStyle} {
-    padding-right: ${({ size }) => sizes[size].height};
-    padding-left: ${({ size }) => sizes[size].horizontalPadding};
-    position: relative;
-
-    ${StyledInput} {
-      padding: 0;
-    }
   }
 `;