+++ src/components/step-sequence/step-sequence.pw.tsx
@@ -7,65 +7,13 @@ import {
 
 import { StepSequenceItemProps } from ".";
 import { checkAccessibility } from "../../../playwright/support/helper";
-import {
-  stepSequenceDataComponent,
-  stepSequenceDataComponentItem,
-} from "../../../playwright/components/step-sequence";
-
-import { ICON } from "../../../playwright/components/locators";
 
 import { CHARACTERS } from "../../../playwright/support/constants";
+import { stepSequenceDataComponentItem } from "../../../playwright/components/step-sequence";
 
 const testData = [CHARACTERS.DIACRITICS, CHARACTERS.SPECIALCHARACTERS];
 
-test.describe("Testing StepSequence component properties", () => {
-  ["horizontal", "vertical"].forEach((orientation) => {
-    test(`should check orientation is set to ${orientation}`, async ({
-      mount,
-      page,
-    }) => {
-      await mount(<StepSequenceComponent orientation={orientation} />);
-
-      await expect(stepSequenceDataComponent(page)).toHaveAttribute(
-        "orientation",
-        orientation,
-      );
-
-      if (orientation === "vertical") {
-        await expect(stepSequenceDataComponent(page)).toHaveCSS(
-          "flex-direction",
-          "column",
-        );
-      }
-    });
-  });
-
-  test("should check StepSequenceItem with children", async ({
-    mount,
-    page,
-  }) => {
-    await mount(<StepSequenceItemCustom />);
-
-    await expect(
-      stepSequenceDataComponentItem(page).locator("span").nth(1),
-    ).toHaveText("Name");
-  });
-
-  ["-100", "0", "999", testData[0], testData[1]].forEach((indicator) => {
-    test(`should check indicator is set to ${indicator}`, async ({
-      mount,
-      page,
-    }) => {
-      await mount(
-        <StepSequenceItemCustom status="incomplete" indicator={indicator} />,
-      );
-      const expectedLabelChild = stepSequenceDataComponentItem(page)
-        .locator("span > span")
-        .nth(0);
-      await expect(expectedLabelChild).toHaveText(indicator);
-    });
-  });
-
+test.describe("tests for StepSequence component", () => {
   testData.forEach((ariaLabel) => {
     test(`should check ariaLabel is set to ${ariaLabel}`, async ({
       mount,
@@ -76,27 +24,6 @@ test.describe("Testing StepSequence component properties", () => {
         "aria-label",
         ariaLabel,
       );
-    });
-  });
-
-  (
-    [
-      ["complete", "rgb(0, 138, 33)"],
-      ["current", "rgba(0, 0, 0, 0.9)"],
-      ["incomplete", "rgba(0, 0, 0, 0.55)"],
-    ] as [StepSequenceItemProps["status"], string][]
-  ).forEach(([status, color]) => {
-    test(`should check status is set to ${status}`, async ({ mount, page }) => {
-      await mount(<StepSequenceItemCustom status={status} />);
-      await expect(stepSequenceDataComponentItem(page)).toHaveCSS(
-        "color",
-        color,
-      );
-
-      const expectedLabelChild = stepSequenceDataComponentItem(page)
-        .locator("span > span")
-        .nth(0);
-      await expect(expectedLabelChild).toHaveCSS("color", color);
     });
   });
 
@@ -136,41 +63,6 @@ test.describe("Testing StepSequence component properties", () => {
       await expect(expectedLabelChild).toHaveText(hiddenCurrentLabel);
     });
   });
-
-  test("should check hideIndicator prop when status is set to complete", async ({
-    mount,
-    page,
-  }) => {
-    await mount(<StepSequenceItemCustom status="complete" hideIndicator />);
-    const expectedLabelChild =
-      stepSequenceDataComponentItem(page).locator(ICON);
-    await expect(expectedLabelChild).toBeVisible();
-    await expect(stepSequenceDataComponentItem(page)).toHaveCSS(
-      "color",
-      "rgb(0, 138, 33)",
-    );
-  });
-
-  (
-    [
-      ["current", "rgba(0, 0, 0, 0.9)"],
-      ["incomplete", "rgba(0, 0, 0, 0.55)"],
-    ] as [StepSequenceItemProps["status"], string][]
-  ).forEach(([status, color]) => {
-    test(`should check hideIndicator prop when status is set to ${status}`, async ({
-      mount,
-      page,
-    }) => {
-      await mount(<StepSequenceItemCustom status={status} hideIndicator />);
-      const expectedLabelChild =
-        stepSequenceDataComponentItem(page).locator(ICON);
-      await expect(expectedLabelChild).toBeHidden();
-      await expect(stepSequenceDataComponentItem(page)).toHaveCSS(
-        "color",
-        color,
-      );
-    });
-  });
 });
 
 test.describe("Accessibility tests for StepSequence component", () => {
@@ -194,26 +86,14 @@ test.describe("Accessibility tests for StepSequence component", () => {
     await checkAccessibility(page);
   });
 
-  ["-100", "0", "999", testData[0], testData[1]].forEach((indicator) => {
-    test(`should check indicator is set to ${indicator}`, async ({
-      mount,
-      page,
-    }) => {
-      await mount(
-        <StepSequenceItemCustom status="incomplete" indicator={indicator} />,
-      );
-      await checkAccessibility(page);
-    });
+  test("should check indicator for accessibility", async ({ mount, page }) => {
+    await mount(<StepSequenceItemCustom status="incomplete" indicator="1" />);
+    await checkAccessibility(page);
   });
 
-  testData.forEach((ariaLabel) => {
-    test(`should check ariaLabel is set to ${ariaLabel}`, async ({
-      mount,
-      page,
-    }) => {
-      await mount(<StepSequenceItemCustom aria-label={ariaLabel} />);
-      await checkAccessibility(page);
-    });
+  test("should check ariaLabel for accessibility", async ({ mount, page }) => {
+    await mount(<StepSequenceItemCustom aria-label="Step 1 of 5" />);
+    await checkAccessibility(page);
   });
 
   (
@@ -225,34 +105,27 @@ test.describe("Accessibility tests for StepSequence component", () => {
     });
   });
 
-  testData.forEach((hiddenCompleteLabel) => {
-    test(`should check hiddenCompleteLabel is set to ${hiddenCompleteLabel}`, async ({
-      mount,
-      page,
-    }) => {
-      await mount(
-        <StepSequenceItemCustom
-          status="complete"
-          hiddenCompleteLabel={hiddenCompleteLabel}
-        />,
-      );
-      await checkAccessibility(page);
-    });
+  test("should check hiddenCompleteLabel for accessibility", async ({
+    mount,
+    page,
+  }) => {
+    await mount(
+      <StepSequenceItemCustom
+        status="complete"
+        hiddenCompleteLabel="Complete"
+      />,
+    );
+    await checkAccessibility(page);
   });
 
-  testData.forEach((hiddenCurrentLabel) => {
-    test(`should check hiddenCurrentLabel is set to ${hiddenCurrentLabel}`, async ({
-      mount,
-      page,
-    }) => {
-      await mount(
-        <StepSequenceItemCustom
-          status="complete"
-          hiddenCurrentLabel={hiddenCurrentLabel}
-        />,
-      );
-      await checkAccessibility(page);
-    });
+  test("should check hiddenCurrentLabel for accessibility", async ({
+    mount,
+    page,
+  }) => {
+    await mount(
+      <StepSequenceItemCustom status="current" hiddenCurrentLabel="Current" />,
+    );
+    await checkAccessibility(page);
   });
 
   (